Show badges on profile, soon on market page
